module setup + relocate existing syntaxes
