+++ src/main/java/org/skriptlang/skript/bukkit/portal/PortalModule.java
@@ -1,0 +1,20 @@
+package org.skriptlang.skript.bukkit.portal;
+
+import java.io.IOException;
+
+import ch.njol.skript.Skript;
+import org.skriptlang.skript.addon.AddonModule;
+import org.skriptlang.skript.addon.SkriptAddon;
+
+public class PortalModule implements AddonModule {
+
+	@Override
+	public void load(SkriptAddon addon) {
+		try {
+			Skript.getAddonInstance().loadClasses("org.skriptlang.skript.bukkit.portal", "elements");
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+}
