+++ src/main/java/org/skriptlang/skript/bukkit/portal/elements/ExprPortalCooldown.java
@@ -1,4 +1,4 @@
-package ch.njol.skript.expressions;
+package org.skriptlang.skript.bukkit.portal.elements;
 
 import ch.njol.skript.classes.Changer.ChangeMode;
 import ch.njol.skript.doc.Description;
